move removeDynamicImports to file that uses it
